sync code

+++ rosen/modules/render_service/core/feature/capture/rs_surface_capture_task_parallel.cpp
@@ -500,8 +500,8 @@ std::unique_ptr<Media::PixelMap> RSSurfaceCaptureTaskParallel::CreatePixelMapByD
     screenCorrection_ = screenManager->GetScreenCorrection(screenId);
     screenRotation_ = node->GetScreenRotation();
     finalRotationAngle_ = CalPixelMapRotation();
-    uint32_t pixmapWidth = node->GetFixedWidth();
-    uint32_t pixmapHeight = node->GetFixedHeight();
+    uint32_t pixmapWidth = static_cast<uint32_t>(node->GetFixedWidth());
+    uint32_t pixmapHeight = static_cast<uint32_t>(node->GetFixedHeight());
     auto bounds = node->GetRenderProperties().GetBoundsGeometry();
     boundsX_ = bounds->GetX();
     boundsY_ = bounds->GetY();
